dtpc: address PR #86 review feedback
Track OutAdu.aggrLength as uvast to match estimateLength() and drop the
now-needless int casts in insertToTopic.

In the bench-dtpc harness, clean up via killm f with ION_NODE_LIST_DIR
exported, shut ION down from an EXIT trap instead of an explicit killm,
and drop the SC1091 suppression.  Remove the ignored "1400 100"
parameters from the tcp protocol lines in both bench.bprc files, and
align the leading whitespace of the dtpcsend usage text.

+++ dtpc/library/libdtpcP.c
@@ -692,7 +692,7 @@ static Object	insertToTopic(unsigned int topicID, Object outAduObj,
 	}
 
 	oK(decodeSdnv(&recordLength, newRecord->length.text));
-	outAdu.aggrLength += (int) recordLength;
+	outAdu.aggrLength += recordLength;
 
 	if (sap->elisionFn != NULL)
 	{
@@ -708,7 +708,7 @@ static Object	insertToTopic(unsigned int topicID, Object outAduObj,
 			return 0;
 		}
 
-		outAdu.aggrLength = (int) newLength;
+		outAdu.aggrLength = newLength;
 	}
 
 	sdr_write(sdr, outAduObj, (char *) &outAdu, sizeof(OutAdu));
